fix: kafka configs

+++ stock-service/src/main/java/com/example/stockservice/kafka/dto/ReserveStockDTO.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ReserveStockDTO {
     private String stockId;
-    private int orderAmount;
+    private long orderAmount;
 }
